requirements.txt
numpy
scipy
scikit-learn
matplotlib
pandas
fastapi
uvicorn
uvloop
httptools
starlette==0.16.0
pydantic==1.8.2
tortoise-orm==0.17.8
passlib==1.7.4
jwt==1.3.1